backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.8.5
python-dotenv==1.0.0
pydantic==2.5.0
qdrant-client==1.8.0
openai==1.12.0
requests==2.31.0
numpy==1.24.3
pydantic==2.5.0
python-multipart==0.0.6